+++ earthfile2llb/interpreter.go
@@ -94,15 +94,6 @@ func (i *Interpreter) Run(ctx context.Context, ef spec.Earthfile) (retErr error)
 	return i.errorf(ef.SourceLocation, "target %s not found", i.target.Target)
 }
 
-func (i *Interpreter) isPipelineTarget(_ context.Context, t spec.Target) bool {
-	for _, stmt := range t.Recipe {
-		if stmt.Command != nil && stmt.Command.Name == "PIPELINE" {
-			return true
-		}
-	}
-	return false
-}
-
 func (i *Interpreter) handleTarget(ctx context.Context, t spec.Target) error {
 	ctx = ContextWithSourceLocation(ctx, t.SourceLocation)
 	// Apply implicit FROM +base
@@ -110,11 +101,6 @@ func (i *Interpreter) handleTarget(ctx context.Context, t spec.Target) error {
 	if err != nil {
 		return i.wrapError(err, t.SourceLocation, "apply FROM")
 	}
-
-	if i.isPipelineTarget(ctx, t) {
-		return i.handlePipelineBlock(ctx, t.Name, t.Recipe)
-	}
-
 	return i.handleBlock(ctx, t.Recipe)
 }
 
@@ -300,8 +286,6 @@ func (i *Interpreter) handleCommand(ctx context.Context, cmd spec.Command) (err 
 		return i.handleHost(ctx, cmd)
 	case command.Project:
 		return i.handleProject(ctx, cmd)
-	case command.Trigger:
-		return i.handleTrigger(ctx, cmd)
 	default:
 		return i.errorf(cmd.SourceLocation, "unexpected command %q", cmd.Name)
 	}
@@ -1987,79 +1971,6 @@ func (i *Interpreter) handleProject(ctx context.Context, cmd spec.Command) error
 	err := i.converter.Project(ctx, parts[0], parts[1])
 	if err != nil {
 		return i.wrapError(err, cmd.SourceLocation, "failed to process PROJECT")
-	}
-
-	return nil
-}
-
-func (i *Interpreter) handlePipelineBlock(ctx context.Context, name string, block spec.Block) error {
-	if len(block) == 0 {
-		return errors.New("pipeline targets require sub-commands")
-	}
-
-	if block[0].Command == nil || block[0].Command.Name != "PIPELINE" {
-		return i.errorf(block[0].Command.SourceLocation, "PIPELINE must be the first command in a pipeline target")
-	}
-
-	for _, stmt := range block {
-		if stmt.Command == nil {
-			return errors.New("pipeline targets do not support IF, WITH, FOR, or WAIT commands")
-		}
-		cmd := *stmt.Command
-		ctx = ContextWithSourceLocation(ctx, cmd.SourceLocation)
-		var err error
-		switch cmd.Name {
-		case command.Pipeline:
-			err = i.handlePipeline(ctx, cmd)
-		case command.Trigger:
-			err = i.handleTrigger(ctx, cmd)
-		case command.Arg:
-			err = i.handleArg(ctx, cmd)
-		case command.Build:
-			err = i.handleBuild(ctx, cmd, false)
-		default:
-			return i.errorf(cmd.SourceLocation, "pipeline targets only support PIPELINE, TRIGGER, ARG, and BUILD commands")
-		}
-		if err != nil {
-			return err
-		}
-	}
-
-	return nil
-}
-
-func (i *Interpreter) handlePipeline(ctx context.Context, cmd spec.Command) error {
-
-	if len(cmd.Args) > 1 {
-		return i.errorf(cmd.SourceLocation, "invalid number of PIPELINE arguments")
-	}
-
-	var opts commandflag.PipelineOpts
-	_, err := flagutil.ParseArgsCleaned("PIPELINE", &opts, flagutil.GetArgsCopy(cmd))
-	if err != nil {
-		return i.wrapError(err, cmd.SourceLocation, "invalid PIPELINE arguments")
-	}
-
-	return i.converter.Pipeline(ctx)
-}
-
-func (i *Interpreter) handleTrigger(ctx context.Context, cmd spec.Command) error {
-
-	if len(cmd.Args) < 1 {
-		return i.errorf(cmd.SourceLocation, "TRIGGER requires at least 1 argument")
-	}
-
-	switch cmd.Args[0] {
-	case "manual":
-		if len(cmd.Args) != 1 {
-			return i.errorf(cmd.SourceLocation, "invalid argument")
-		}
-	case "pr", "push":
-		if len(cmd.Args) != 2 {
-			return i.errorf(cmd.SourceLocation, "'pr' and 'push' triggers require a branch name")
-		}
-	default:
-		return i.errorf(cmd.SourceLocation, "valid triggers include: 'manual', 'pr', or 'push'")
 	}
 
 	return nil